make_posts/tests/factories.py
import factory
from factory.django import DjangoModelFactory
from make_posts.models import User


class UserFactory(DjangoModelFactory):
    class Meta:
        model = User

    username = factory.Faker('user_name')
    password = factory.Faker('password')
